changed eslintrc file for single quotes

+++ mashup/.eslintrc.js
@@ -11,7 +11,7 @@ module.exports = {
     sourceType: "module"
   },
   rules: {
-    quotes: [2, "double", { avoidEscape: true }],
+    quotes: [2, "single", { avoidEscape: true }],
     semi: ["error", "always"]
   }
 };
